// FamilyClub.React/famiy.club.react/lib/api/generated/models/PublisherDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PublisherDto
 */
export interface PublisherDto {
    /**
     * 
     * @type {number}
     * @memberof PublisherDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PublisherDto
     */
    publisherName?: string | null;
}

/**
 * Check if a given object implements the PublisherDto interface.
 */
export function instanceOfPublisherDto(value: object): value is PublisherDto {
    return true;
}

export function PublisherDtoFromJSON(json: any): PublisherDto {
    return PublisherDtoFromJSONTyped(json, false);
}

export function PublisherDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PublisherDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'publisherName': json['publisherName'] == null ? undefined : json['publisherName'],
    };
}

export function PublisherDtoToJSON(json: any): PublisherDto {
    return PublisherDtoToJSONTyped(json, false);
}

export function PublisherDtoToJSONTyped(value?: PublisherDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'publisherName': value['publisherName'],
    };
}